Move functions to qc_utils
